[chore]: 시퀀스 스토어 관련 커스텀 훅 추가 및 기존 모델 파일 삭제
