+++ src/widgets/scenario-list/ui/icon/StatusIcon.tsx
@@ -1,4 +1,4 @@
-import { IconType } from "@/entities/sequence/model/sequence-store";
+import { IconType } from "@/entities/sequence/types/sequence";
 import { vars } from "@/shared/ui/styles/theme.css";
 import { FiLoader as LoadingIcon } from "react-icons/fi";
 import { IoCheckmarkCircleSharp as CheckIcon } from "react-icons/io5";
