+++ src/widgets/sequence-list/ui/item/SequenceItem.tsx
@@ -1,7 +1,7 @@
 import { vars } from "@/shared/ui/styles/theme.css";
 import classNames from "classnames";
 
-import type { SequenceItemType } from "@/entities/sequence/model/sequence-store";
+import type { SequenceItemType } from "@/entities/sequence/types/sequence";
 import { sequenceItemStyles } from "./sequenceItem.css";
 
 export interface SequenceItemProps extends SequenceItemType {
